requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
redminelib>=3.0.0
requests>=2.28.0
openpyxl>=3.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0 